requirements .txt
aiofiles
aioshutil
anytree
apscheduler
aria2p
asyncio
cloudscraper
dnspython
feedparser
flask
gevent
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
gunicorn
httpx
lxml
motor
natsort
par2cmdline-turbo
pillow
psutil
pymongo
pyrofork